+++ converter/converters/affinity/koki_affinity_to_kube_v1_affinity_test.go
@@ -1,4 +1,4 @@
-package converters
+package affinity
 
 import (
 	"testing"
@@ -77,7 +77,7 @@ func TestRevert(t *testing.T) {
 }
 
 func testRevertAffinities(t *testing.T, as ...types.Affinity) {
-	_, err := convert_Koki_Affinity_to_Kube_v1_Affinity(as)
+	_, err := Convert_Koki_Affinity_to_Kube_v1_Affinity(as)
 	if err != nil {
 		t.Error(err)
 	}
